regular maintenance

+++ Backend/routes/server.ts
@@ -3,8 +3,15 @@ import prisma from "../db";
 
 const app = Router();
 
-app.get("/wake-up", async (req, res) => {
+import * as dotenv from "dotenv";
+dotenv.config();
+
+app.get("/wake-up/:pwd", async (req, res) => {
   try {
+    const pwd = req.params.pwd;
+    if (pwd !== process.env.RESET_PWD) {
+      throw new Error("Invalid Password");
+    }
     const users = await prisma.user.findMany();
     return res.json({ success: 1, users });
   } catch (error: any) {
@@ -13,8 +20,12 @@ app.get("/wake-up", async (req, res) => {
   }
 });
 
-app.get("/reset", async (req, res) => {
+app.get("/reset/:pwd", async (req, res) => {
   try {
+    const pwd = req.params.pwd;
+    if (pwd !== process.env.RESET_PWD) {
+      throw new Error("Invalid Password");
+    }
     await prisma.user.deleteMany();
     await prisma.files.deleteMany();
     return res.send({ success: 1 });
